+++ src/curvesimilarities/averagefrechet.py
@@ -1,16 +1,8 @@
 """Average Fréchet distance and its variants."""
 
 import numpy as np
-from numba import njit
 
-from .integfrechet import (
-    _ifd,
-    _ifd_owp,
-    _line_point_integrate_SqEuc,
-    _sample_ifd_pts,
-    sanitize_vertices_ifd,
-)
-from .util import sanitize_vertices
+from .integfrechet import ifd, ifd_owp
 
 __all__ = [
     "qafd",
@@ -18,20 +10,7 @@ __all__ = [
 ]
 
 
-@njit(cache=True)
-def qafd_degenerate(curve, point):
-    ret = 0
-    length = 0
-    for i in range(len(curve) - 1):
-        a, b = curve[i], curve[i + 1]
-        ret += _line_point_integrate_SqEuc(a, b, point)
-        length += np.linalg.norm(b - a)
-    return np.sqrt(ret / length)
-
-
-@sanitize_vertices(owp=False)
-@sanitize_vertices_ifd(qafd_degenerate, owp=False)
-def qafd(P, Q, delta):
+def qafd(P, Q, delta, dist="euclidean"):
     r"""Quadratic average Fréchet distance between two open polygonal curves.
 
     The quadratic average Fréchet distance is defined as
@@ -42,19 +21,19 @@ def qafd(P, Q, delta):
         \sqrt{
             \frac{
                 \int_0^1
-                \delta\left(\pi(t)\right)^2 \cdot
-                \lVert \pi'(t) \rVert
+                dist\left(\pi(t)\right)^2 \cdot
+                \lVert \pi'(t) \rVert_1
                 \mathrm{d}t
             }{
                 \int_0^1
-                \lVert \pi'(t) \rVert
+                \lVert \pi'(t) \rVert_1
                 \mathrm{d}t
             }
         },
 
-    where :math:`\delta` is the Euclidean distance and :math:`\pi` is the continuous
-    nondecreasing path in the parameter space. We use the Manhattan norm for
-    :math:`\lVert \cdot \rVert`, so the formula can be reduced to
+    where :math:`dist` is the Euclidean distance and :math:`\pi` is the continuous
+    nondecreasing path in the parameter space. Because we use the Manhattan norm
+    :math:`\lVert \cdot \rVert_1`, the formula can be reduced to
 
     .. math::
 
@@ -62,8 +41,8 @@ def qafd(P, Q, delta):
         \inf_{\pi}
         \sqrt{
         \int_0^1
-        \delta\left(\pi(t)\right)^2 \cdot
-        \lVert \pi'(t) \rVert
+        dist\left(\pi(t)\right)^2 \cdot
+        \lVert \pi'(t) \rVert_1
         \mathrm{d}t
         },
 
@@ -79,10 +58,12 @@ def qafd(P, Q, delta):
         :math:`n`-dimensional space.
     delta : double
         Maximum length of edges between Steiner points.
+    dist : {'euclidean'}
+        Type of :math:`dist`. Refer to the Notes section for more information.
 
     Returns
     -------
-    dist : double
+    double
         The quadratic average Fréchet distance between *P* and *Q*, NaN if any
         vertice is empty or both vertices consist of a single point.
 
@@ -101,28 +82,16 @@ def qafd(P, Q, delta):
     >>> qafd([[0, 0], [0.5, 0], [1, 0]], [[0, 1], [1, 1]], 0.1)
     1.0
     """
-    (
-        P_edge_len,
-        P_subedges_num,
-        P_pts,
-        Q_edge_len,
-        Q_subedges_num,
-        Q_pts,
-    ) = _sample_ifd_pts(P, Q, delta)
-    square_ifd = _ifd(
-        P_edge_len,
-        P_subedges_num,
-        P_pts,
-        Q_edge_len,
-        Q_subedges_num,
-        Q_pts,
-    )
-    return float(np.sqrt(square_ifd / (np.sum(P_edge_len) + np.sum(Q_edge_len))))
+    if dist == "euclidean":
+        square_ifd = ifd(P, Q, delta, dist="squared_euclidean")
+    else:
+        raise ValueError(f"Unknown type of distance: {dist}")
+    P_len = np.sum(np.linalg.norm(np.diff(P, axis=0), axis=-1))
+    Q_len = np.sum(np.linalg.norm(np.diff(Q, axis=0), axis=-1))
+    return float(np.sqrt(square_ifd / (P_len + Q_len)))
 
 
-@sanitize_vertices(owp=True)
-@sanitize_vertices_ifd(qafd_degenerate, owp=True)
-def qafd_owp(P, Q, delta):
+def qafd_owp(P, Q, delta, dist="euclidean"):
     """Quadratic average Fréchet distance and its optimal warping path.
 
     Parameters
@@ -135,6 +104,8 @@ def qafd_owp(P, Q, delta):
         :math:`n`-dimensional space.
     delta : double
         Maximum length of edges between Steiner points.
+    dist : {'euclidean'}
+        Type of :math:`dist`. Refer to :func:`qafd`.
 
     Returns
     -------
@@ -156,21 +127,8 @@ def qafd_owp(P, Q, delta):
     >>> import matplotlib.pyplot as plt #doctest: +SKIP
     >>> plt.plot(*path.T)  #doctest: +SKIP
     """
-    (
-        P_edge_len,
-        P_subedges_num,
-        P_pts,
-        Q_edge_len,
-        Q_subedges_num,
-        Q_pts,
-    ) = _sample_ifd_pts(P, Q, delta)
-    dist, path, count = _ifd_owp(
-        P_edge_len,
-        P_subedges_num,
-        P_pts,
-        Q_edge_len,
-        Q_subedges_num,
-        Q_pts,
-    )
-    owp = path[:count]
+    if dist == "euclidean":
+        dist, owp = ifd_owp(P, Q, delta, dist="squared_euclidean")
+    else:
+        raise ValueError(f"Unknown type of distance: {dist}")
     return float(np.sqrt(dist / np.sum(owp[-1]))), owp